backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
python-multipart==0.0.9
# Note: ChromaDB may require Visual C++ Build Tools on Windows
# If installation fails, see README_WINDOWS.md
chromadb==0.5.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-core==0.3.0
tiktoken==0.9.0
pydantic==2.9.0